+++ src/Constraints/Database/ForeignKey.php
@@ -7,8 +7,6 @@ use TanoConsulting\DataValidatorBundle\Constraints\DatabaseConstraint;
 /// @todo in constructor validate contents of members from/to/except
 class ForeignKey extends DatabaseConstraint
 {
-    static protected $defaultName = 'foreign_key_';
-
     public $child;
 
     public $parent;