+++ src/Constraints/Filesystem/DirectoryConstraint.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace TanoConsulting\DataValidatorBundle\Constraints\Filesystem;
+
+use TanoConsulting\DataValidatorBundle\Constraints\FilesystemConstraint;
+
+class DirectoryConstraint extends FilesystemConstraint
+{
+}