+++ src/Mapping/FilesystemMetaData.php
@@ -22,7 +22,7 @@ class FilesystemMetaData extends Metadata
     protected function checkConstraint(Constraint $constraint)
     {
         $targets = (array) $constraint->getTargets();
-        if (!\in_array(FilesystemConstraint::DIRECTORY_CONSTRAINT, $targets, true) && !\in_array(FilesystemConstraint::FILE_CONSTRAINT, $targets, true)) {
+        if (!\in_array(FilesystemConstraint::FILESYSTEM_CONSTRAINT, $targets, true)) {
             throw new ConstraintDefinitionException(sprintf('The constraint "%s" cannot be put on filesystems.', get_debug_type($constraint)));
         }